--- src/constants/calendar/scheduleType.js
@@ -1,5 +1,0 @@
-const SCHEDULE_TYPE = {
-	SHARED: "SHARE_SCHEDULE",
-	PERSONAL: "PERSONAL_SCHEDULE",
-};
-export default SCHEDULE_TYPE;
